Trinity.Core/Trinity.C: TrinityMessageType width fixed -- should be uint16 not uint8. This also fixes non-deterministic failures of native message handlers -- the TrinityMessageType header was not filled completely

+++ src/Trinity.C/src/Network/ProtocolConstants.h
@@ -12,7 +12,7 @@ namespace Trinity
     {
         typedef long double float_t;
 
-        enum TrinityMessageType : uint8_t
+        enum TrinityMessageType : uint16_t
         {
             SYNC,
             SYNC_WITH_RSP,
